Program is now able to build a pre-calculated set of frequencies which are then used in an input file by the program. This allows more spectral densities to be used (ie those that cannot be integrated analytically)
Note that :
a) The slowdown caused by the many reads from file of the frequencies is significant, adding around 0.3 seconds per timestep to running (or 30 minutes to a run of 6000 timesteps)
b) sub-ohmic spectral densities (ie those with s<1) still pose a problem for the integrator.

As the frequencies which are being read (every time the hamiltonian is calculated and every time the differential of the hamiltonian is calculated, so multiple times per time step) do not change over the entire propagation, it may be possible to read only once and carry these values over. This would be difficult due to the layout of the program itself.